sample-repos/python-django/requirements.txt
# Django web framework
Django==4.2.0
psycopg2-binary==2.9.9
celery==5.3.0
redis==5.0.0

# Known vulnerability for testing (CVE-2023-32681)
requests==2.25.0

# API and utilities
djangorestframework==3.14.0
django-cors-headers==4.3.0
gunicorn==21.2.0

# Testing
pytest==7.4.0
pytest-django==4.5.0
